% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfidenceInterval.R
\name{getConfidenceInterval}
\alias{getConfidenceInterval}
\title{compute 95% confidence interval (frequentist approach)
assume the input samples follow normal distribution}
\usage{
getConfidenceInterval(vector)
}
\arguments{
\item{vector}{a numeric vector}
}
\value{
vector a numeric vector with two elements only
}
\description{
compute 95% confidence interval (frequentist approach)
assume the input samples follow normal distribution
}
\examples{

#example with toy samples
sample<-c(1,2,3,4,3,2,6)
getConfidenceInterval(sample)

}
