% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCredibleInterval.R
\name{getCredibleInterval}
\alias{getCredibleInterval}
\title{compute 95% credible interval (bayesian approach)
assume the input prior distribution and samples follow normal distribution}
\usage{
getCredibleInterval(vector, prior_dis, sample_dis)
}
\arguments{
\item{vector}{a numeric vector}

\item{prior_dis}{a numeric vector}

\item{smaple_dis}{a numeric vector}
}
\value{
vector a numeric vector with two elements only
}
\description{
compute 95% credible interval (bayesian approach)
assume the input prior distribution and samples follow normal distribution
}
\examples{

#example with toy samples
sample<-rnorm(5,mean=1,sd=0.2)
getCredibleInterval(sample,c(1.4,0.3),c(1,0.2))

}
